Migrate Editor prompt
